update quests cache more often
