+++ app/perftools/management/commands/create_page_cache.py
@@ -192,25 +192,6 @@ def create_grants_cache():
         )
 
 
-def create_quests_cache():
-    from quests.helpers import generate_leaderboard
-    from quests.views import current_round_number
-    for i in range(1, current_round_number+1):
-        print(f'quests_{i}')
-        view = 'quests'
-        keyword = f'leaderboard_{i}'
-        data = generate_leaderboard(round_number=i)
-        JSONStore.objects.create(
-            view=view,
-            key=keyword,
-            data=json.loads(json.dumps(data, cls=EncodeAnything)),
-            )
-
-    from quests.models import Quest
-    for quest in Quest.objects.filter(visible=True):
-        quest.save()
-
-
 def create_results_cache():
     print('results')
     keywords = ['']
@@ -268,6 +249,5 @@ class Command(BaseCommand):
             create_post_cache()
             create_results_cache()
             create_avatar_cache()
-            create_quests_cache()
             create_grants_cache()
             create_contributor_landing_page_context()
